requirements.txt
gunicorn
django-filter>=2.4.0
djangorestframework>=3.8.2
Markdown
django-cors-headers>=3.7.0
dj-database-url>=0.5.0
django-ses
django>=3.1.0
psycopg2
celery
djangorestframework-simplejwt==4.7.2
raven
django-sentry-400-middleware
ipython
coverage
flake8
flake8-docstrings
pytest
pytest-xdist
pytest-cov
pytest-django
model-mommy
coinoxr
tox
pep257
ujson
autopep8
